Create wallet management-service; Complete communication with payment and wallet service

+++ Backend/payment-service/src/main/java/com/forex/backend/payment_service/dto/TransactionExecutionDTO.java
@@ -2,7 +2,8 @@ package com.forex.backend.payment_service.dto;
 
 import lombok.Builder;
 
+// This dto sends required data to initiate the creation of intent in stripe and execute transaction
 @Builder
-public record TransactionExecutionDTO (Long amount, String paymentMethodId, String idempotencyKey, String currency,
+public record TransactionExecutionDTO (Double amount, String paymentMethodId, String idempotencyKey, String currency,
                                        String customerEmail, String transactionId) {
 }
